Rename package and project to "nicknames"
