+++ python/hatch_build.py
@@ -31,4 +31,4 @@ from hatchling.builders.hooks.plugin.interface import BuildHookInterface
 
 class CustomHook(BuildHookInterface):
     def initialize(self, version, build_data):
-        shutil.copyfile("../names.csv", "./src/us_nicknames/names.csv")
+        shutil.copyfile("../names.csv", "./src/nicknames/names.csv")
