+++ python/src/nicknames/_version.py
